+++ extra_platforms/__init__.py
@@ -399,7 +399,7 @@ Pytest optional.
 """
 
 
-__version__ = "12.0.1"
+__version__ = "12.0.2.dev0"
 
 
 def _initialize_group_detection_functions() -> list[str]: